Pass explicit argvs to exec calls, avoiding shells
Shells are dangerous to use, and we should avoid them whenever possible.
We're all adults here, but a future change might not be inspected with
enough care and cause dire problems. Interacting with the internet and
unzipping internetdatastuff near shells interpretation is deeply spooky.
Let's just be safe, instead.

Incidentally this was double as slow as it had to be, because the "wget"
line was downloading AllPublicXML.zip twice! Mistakes happen.