Make `previous_state` reflect what a user expects
In the case where a trial has been seen to be overdue for the first
time, but on the very same day that we've first seen a QA interaction
that might cancel the overdue status, then reset the `status` to
`ongoing`.

This is because we use `previous_status` to mean "status that could
have been seen on previous days"; when two state changes originate
from two different sources (the ClincialTrials download, and scraping
the QA tabs) on the same day, it is (in our current workflow)
impossible for the first state change to have been seen.

This unpleasant hack is a consequence of the fact our data model
really needs refactoring - see the README for context