primary_purpose update
Updates ACT/pACT evaluation to include instances where the `primary_purpose` field is null, as this was not previously being considered with the `<>` operator. On a quick investigation, this seems to only impact pACTs at the moment and should add ~5 overdue trials to the tracker tomorrow.

As a check, our other field using as `<>` operator (`study_status`) is never null so this is not an issue for that field.