Fixes has_results issue
Currently the `has_results` field uses the presence of the `clinical_results` section of the JSON to determine if results are available for viewing. This is breaking everything because one trial submitted an enormous results section. We can use the availability of the field `results_first_submitted` instead as this will never appear unless there are results, and it's just a date.