New pACT criteria
This expands the pACT criteria to include trials that are not explicitly conducted in the US but have updated one of the new `Studies a U.S. FDA-regulated Drug Product` or `Studies a U.S. FDA-regulated Device Product` fields to `Yes`.

If one of these fields is `Yes` it tells us that the trial was either conducted in the US, conducted under an IND/IDE or conducted on a product that was manufactured and exported from the US. We know that this field cannot be `Yes` unless one of those things is true and currently use the `FDA-Regulated` fields in this way when making ACT assessments based on feedback from ClinicalTrials.gov staff. 

This change extends that logic to pACTs that have explicitly updated these fields. These fields are not currently required for trials that began pre-Final Rule but they are often updated anyway during routine updates to registry information. Trials that have not updated these fields will be held to the same pACT status as before (including an explicit US location) as outlined in the Clinicaltrials.gov PRS UserGuide.

More details on this change will be available in an update to our methods pre-print paper in the near future.