Update to align with new SQL
This should update the python to align with the new SQL. It might be a bit before we get this reviewed and deployed in the pipeline, but we need to make the new pACT identification changes sooner. This should bring those two into harmony so we can test and compare the datasets for consistency in the future before integrating.