Updated to address comments
Updated the script with feedback from @sebbacon and @chadmiller.

One note that might look weird to anyone reviewing, I've interspersed the `dict_builder` function on what appears to be a random order throughout the loop, but this is just so the CSV output matches the old CSV format. Just makes my life easier to compare. Once this code goes live, it won't matter what order the CSV is in. Once everything is tested, we can adjust if it is really driving anyone crazy. Can also group them all at the bottom of the loop.

Updates include:
-Replacing use of Pandas with a dictionary
-Created a whole host of helper functions to cut down redundant lines
-Created a dictionary of all the trial data. The Keys are the column headers, and the value for each header is a list of the values for each trial. The CSV rows are built from these lists.
-Removed loads (and learned my lesson about) unnecessary newlines
-Various other small fixes and enhancements throughout.

I assume there's a more efficient way to build my dictionary, rather than repeating my `dict_builder` function over and over again but that's the only way I could get it to work with `defaultdict(list)`.