Drop computable Rankings.percentage and provide func
At model instance .save() time, we were calculating the value of one column
using two others.

Since the type of that column is an IntegerField but elsewhere on the site we
round to one digit after decimal, a reader complained about the inconsistency
(really that it was a floor() operation).

Instead, this is the first part of the whole fix, but someone else can
"makemigrations" to discard the db column at leisure.

Closes #184