Fixing the QA process
This is the first step to fix the QA checking process. Eventually we can probably remove this entirely but for now, this will fix it

Tied to #198

What I've done here is go looking for the text in the "Results" tab. If it says "Study Results" that means that results have been posted. Recently we've been running into and issue where the XMLs have not yet been updated to reflect results being posted, but the live website has results posted. When this code goes looking for the results submitted box, it can't find it and thinks the results disappeared.

I've added another condition to this function that will check for the posting of results (The tab will always read 'Study Results' in that instance. I don't know exactly what to do next, but basically we just need to tell the function to not do anything with that trial if it finds that condition. By the next data refresh the results will likely be posted and then it won't go through the QA check process anyway.