Updating QA trials check pipeline
This replaces the old `set_qa_metadata` function.

The old function had become deprecated with the move to the new version of ClinicalTrials.gov.

Instead of scraping, we now get the relevant data directly from the ClinicalTrials.gov API which simplifies things and should ideally make the pipeline more robust and less prone to failure.